+++ src/GitHub_Updater/Theme.php
@@ -100,13 +100,13 @@ class Theme extends Base {
 	protected function get_theme_meta() {
 		$git_themes = array();
 		$themes     = wp_get_themes( array( 'errors' => null ) );
+		$additions  = apply_filters( 'github_updater_additions', null, $themes, 'theme' );
 
 		foreach ( (array) $themes as $theme ) {
 			$git_theme           = array();
 			$repo_uri            = null;
 			$repo_enterprise_uri = null;
 			$repo_enterprise_api = null;
-			$additions           = apply_filters( 'github_updater_additions', null, $theme, 'theme' );
 
 			foreach ( (array) self::$extra_headers as $value ) {
 
@@ -361,6 +361,10 @@ class Theme extends Base {
 	 * @author Seth Carstens
 	 */
 	public function wp_theme_update_row( $theme_key, $theme ) {
+		global $wp_version;
+		$open_div  = '<div class="update-message">';
+		$close_div = '';
+
 		$current            = get_site_transient( 'update_themes' );
 		$themes_allowedtags = array(
 			'a'       => array( 'href' => array(), 'title' => array() ),
@@ -383,10 +387,23 @@ class Theme extends Base {
 			),
 			$install_url ) );
 
+		/*
+		 * Update transient if necessary.
+		 */
+		if ( empty( $current->response ) && empty( $current->up_to_date ) ) {
+			$this->pre_set_site_transient_update_themes( $current );
+		}
+
+		if ( version_compare( $wp_version, '4.6', '>=' ) || class_exists( 'Shiny_Updates' ) ) {
+			$open_div  = '<div class="update-message notice inline notice-warning notice-alt"><p>';
+			$close_div = '</p>';
+		}
+
+		echo '<tr class="plugin-update-tr" data-slug="' . $theme_key . '"><td colspan="' . $wp_list_table->get_column_count() . '" class="plugin-update colspanchange">' . $open_div;
+
 		if ( isset( $current->up_to_date[ $theme_key ] ) ) {
 			$rollback      = $current->up_to_date[ $theme_key ]['rollback'];
 			$rollback_keys = array_keys( $rollback );
-			echo '<tr class="plugin-update-tr" data-slug="' . $theme_key . '"><td colspan="' . $wp_list_table->get_column_count() . '" class="plugin-update colspanchange"><div class="update-message update-ok">';
 			esc_html_e( 'Theme is up-to-date!', 'github-updater' );
 			echo '&nbsp';
 			if ( count( $rollback ) > 0 ) {
@@ -416,15 +433,14 @@ class Theme extends Base {
 
 		if ( isset( $current->response[ $theme_key ] ) ) {
 			$r = $current->response[ $theme_key ];
-			echo '<tr class="plugin-update-tr" data-slug="' . $theme_key . '"><td colspan="' . $wp_list_table->get_column_count() . '" class="plugin-update colspanchange"><div class="update-message">';
+			printf( esc_html__( 'GitHub Updater shows a new version of %s available.', 'github-updater' ),
+				$theme_name
+			);
+			printf( ' <a href="%s" class="thickbox" title="%s"> ',
+				$details_url,
+				$theme_name
+			);
 			if ( empty( $r['package'] ) ) {
-				printf( esc_html__( 'GitHub Updater shows a new version of %s available.', 'github-updater' ),
-					$theme_name
-				);
-				printf( ' <a href="%s" class="thickbox" title="%s"> ',
-					$details_url,
-					$theme_name
-				);
 				printf( esc_html__( 'View version %s details.', 'github-updater' ),
 					$r['new_version']
 				);
@@ -432,13 +448,6 @@ class Theme extends Base {
 				esc_html_e( 'Automatic update is unavailable for this theme.', 'github-updater' );
 				echo '</em>';
 			} else {
-				printf( esc_html__( 'GitHub Updater shows a new version of %s available.', 'github-updater' ),
-					$theme_name
-				);
-				printf( ' <a href="%s" class="thickbox" title="%s"> ',
-					$details_url,
-					$theme_name
-				);
 				printf( esc_html__( 'View version %1$s details%2$s or %3$supdate now%4$s.', 'github-updater' ),
 					$r['new_version'],
 					'</a>',
@@ -449,7 +458,7 @@ class Theme extends Base {
 
 			do_action( "in_theme_update_message-$theme_key", $theme, $r );
 		}
-		echo '</div></td></tr>';
+		echo $close_div . '</div></td></tr>';
 	}
 
 	/**